accept new solution
